# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: WhatsOpí API
  description: |
    **AI-Powered Digital Platform for Dominican Republic's Informal Economy**
    
    The WhatsOpí API provides comprehensive financial and commerce services specifically designed for the Dominican Republic's informal economy. Built with cultural awareness, offline-first architecture, and optimized for Caribbean networks.
    
    ## Features
    - **Dominican Spanish & Haitian Creole Support**: Native language processing
    - **WhatsApp Integration**: Primary communication channel
    - **Voice Interface**: Caribbean accent-optimized speech recognition
    - **Local Payments**: tPago, PayPal, cards, and cash-in/cash-out via colmados
    - **Offline-First**: PWA capabilities with background sync
    - **Cultural Intelligence**: AI that understands Dominican context
    
  version: 1.0.0
  contact:
    name: WhatsOpí API Support
    email: api-support@whatsopi.do
    url: https://docs.whatsopi.do
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://terms.whatsopi.do
servers:
  - url: https://api.whatsopi.do
    description: Production server
  - url: https://staging-api.whatsopi.do
    description: Staging server
  - url: https://sandbox-api.whatsopi.do
    description: Sandbox server for testing
paths:
...
  /api/v1/whatsapp/send:
    post:
      tags:
        - WhatsApp
      summary: Send WhatsApp message
      description: Send messages via WhatsApp Business API
      operationId: sendWhatsAppMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WhatsAppMessageRequest'
      responses:
        '200':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WhatsAppMessageResponse'

  # Health Check
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoints
  schemas:
...
    WhatsAppMessageRequest:
      type: object
      properties:
        to:
          type: string
          pattern: '^(\+1)?(809|829|849)[0-9]{7}$'
          example: "+18091234567"
        type:
          type: string
          enum: [template, text, interactive]
        template:
          type: object
          description: WhatsApp template message
        text:
          type: string
          description: Plain text message
        interactive:
          type: object
          description: Interactive message with buttons/lists
      required:
        - to
        - type

    # Response Schemas
...
    WhatsAppMessageResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                messageId:
                  type: string
                  example: "wamid.abc123"
                status:
                  type: string
                  example: "sent"
                to:
                  type: string
                  example: "+18091234567"

    # Entity Schemas
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        timestamp:
          type: string
          format: date-time
          example: "2024-12-01T10:00:00Z"
...